kafka-producer/requirements.txt
kafka-python
requests
python-dotenv
boto3
psycopg2-binary
